+++ internal/database/003_strava_tokens.sql.go
@@ -89,28 +89,28 @@ WHERE su.user_id = $1
 `
 
 type GetStravaAccessTokensByApplicationUserIDRow struct {
-	ID           uuid.UUID      `json:"id"`
-	StravaUserID uuid.UUID      `json:"strava_user_id"`
-	AccessToken  string         `json:"access_token"`
-	TokenType    string         `json:"token_type"`
-	ExpiresAt    time.Time      `json:"expires_at"`
-	RefreshToken string         `json:"refresh_token"`
-	CreatedAt    time.Time      `json:"created_at"`
-	UpdatedAt    time.Time      `json:"updated_at"`
-	ID_2         uuid.UUID      `json:"id_2"`
-	UserID       uuid.UUID      `json:"user_id"`
-	StravaID     int64          `json:"strava_id"`
-	CreatedAt_2  time.Time      `json:"created_at_2"`
-	UpdatedAt_2  time.Time      `json:"updated_at_2"`
-	Username     string         `json:"username"`
-	Firstname    sql.NullString `json:"firstname"`
-	Lastname     sql.NullString `json:"lastname"`
-	City         sql.NullString `json:"city"`
-	State        sql.NullString `json:"state"`
-	Country      sql.NullString `json:"country"`
-	Sex          sql.NullString `json:"sex"`
-	Premuim      sql.NullBool   `json:"premuim"`
-	Weight       sql.NullInt32  `json:"weight"`
+	ID           uuid.UUID       `json:"id"`
+	StravaUserID uuid.UUID       `json:"strava_user_id"`
+	AccessToken  string          `json:"access_token"`
+	TokenType    string          `json:"token_type"`
+	ExpiresAt    time.Time       `json:"expires_at"`
+	RefreshToken string          `json:"refresh_token"`
+	CreatedAt    time.Time       `json:"created_at"`
+	UpdatedAt    time.Time       `json:"updated_at"`
+	ID_2         uuid.UUID       `json:"id_2"`
+	UserID       uuid.UUID       `json:"user_id"`
+	StravaID     int64           `json:"strava_id"`
+	CreatedAt_2  time.Time       `json:"created_at_2"`
+	UpdatedAt_2  time.Time       `json:"updated_at_2"`
+	Username     string          `json:"username"`
+	Firstname    sql.NullString  `json:"firstname"`
+	Lastname     sql.NullString  `json:"lastname"`
+	City         sql.NullString  `json:"city"`
+	State        sql.NullString  `json:"state"`
+	Country      sql.NullString  `json:"country"`
+	Sex          sql.NullString  `json:"sex"`
+	Premuim      sql.NullBool    `json:"premuim"`
+	Weight       sql.NullFloat64 `json:"weight"`
 }
 
 func (q *Queries) GetStravaAccessTokensByApplicationUserID(ctx context.Context, userID uuid.UUID) (GetStravaAccessTokensByApplicationUserIDRow, error) {
